+++ REmatchEngine/src/match.hpp
@@ -22,7 +22,7 @@ using Match_ptr = unique_ptr<Match>;
 #endif
 
 
-using Span = pair<size_t, size_t>;
+using Span = pair<int64_t, int64_t>;
 using SpanMap = map<string, Span>;
 using SpanVect = vector<Span>;
 
@@ -40,8 +40,8 @@ class Match {
   Match(string const &d, SpanVect s,
         vector<string> output_scheme);
 
-  size_t start(string varname) const {return span(varname).first;}
-  size_t end(string varname) const {return span(varname).second;}
+  int64_t start(string varname) const {return span(varname).first;}
+  int64_t end(string varname) const {return span(varname).second;}
 
   // Returns a pair<uint64_t, uint64_t> correspoding to a variable's span.
   Span span(string var) const;
